+++ pairing-tool/index.js
@@ -3,7 +3,20 @@ const flicLib = require('fliclib-linux-hci')
 const fs = require('fs')
 const CsvWriter = require('csv-write-stream')
 const { prompt } = require('enquirer')
+const dotenv = require('dotenv')
+const twilio = require('twilio')
 
+// Setup environment variables
+dotenv.config()
+const DOMAIN = process.env.DOMAIN
+const TWILIO_MESSAGING_SERVICE_ID = process.env.TWILIO_MESSAGING_SERVICE_ID
+
+// Setup Twilio
+const TWILIO_SID = process.env.TWILIO_SID
+const TWILIO_TOKEN = process.env.TWILIO_TOKEN
+const twilioClient = twilio(TWILIO_SID, TWILIO_TOKEN)
+
+// Setup Flic
 const FlicClient = flicLib.FlicClient
 const FlicScanWizard = flicLib.FlicScanWizard
 
@@ -26,6 +39,12 @@ async function scanAndPairOneButton() {
       if (result !== 'WizardSuccess') {
         resolve()
       } else {
+        const clientQuestion = {
+          type: 'input',
+          name: 'clientName',
+          message: `What is this new button's client's name?`,
+        }
+
         const unitQuestion = {
           type: 'input',
           name: 'unit',
@@ -34,8 +53,8 @@ async function scanAndPairOneButton() {
 
         const phoneNumberQuestion = {
           type: 'input',
-          name: 'phoneNumber',
-          message: 'What is the phone number for this new button?',
+          name: 'phoneNumberOrAreaCode',
+          message: 'What is the phone number for this new button (or enter an area code to buy a new phone number)?',
         }
 
         const confirmQuestion = {
@@ -44,12 +63,51 @@ async function scanAndPairOneButton() {
           message: 'Create a button with the info above?',
         }
 
-        const response = await prompt([unitQuestion, phoneNumberQuestion, confirmQuestion])
+        const response = await prompt([clientQuestion, unitQuestion, phoneNumberQuestion, confirmQuestion])
 
         if (response.confirmation) {
           // eslint-disable-next-line no-unused-vars -- these vars may be useful soon
-          client.getButtonInfo(bdAddr, (_bdAddr, uuid, color, serialNumber) => {
-            csvWriter.write([uuid, response.unit, response.phoneNumber, serialNumber])
+          client.getButtonInfo(bdAddr, async (_bdAddr, uuid, color, serialNumber) => {
+            const phoneNumberOrAreaCode = response.phoneNumberOrAreaCode.trim()
+            let phoneNumber
+            let phoneNumberSid
+            if (phoneNumberOrAreaCode.length === 3) {
+              try {
+                // Buy Twilio number
+                const incomingPhoneNumber = await twilioClient.incomingPhoneNumbers.create({
+                  areaCode: phoneNumberOrAreaCode,
+                  smsUrl: `https://${DOMAIN}/alert/sms`,
+                  voiceUrl: 'https://demo.twilio.com/welcome/voice/',
+                  friendlyName: `${response.clientName.trim()} ${response.unit.trim()}`,
+                })
+                phoneNumber = incomingPhoneNumber.phoneNumber
+                phoneNumberSid = incomingPhoneNumber.sid
+                console.log(`  ... Bought phone number: ${phoneNumber}`)
+              } catch (e) {
+                console.error(
+                  '  ... ERROR buying new Twilio phone number, check for Twilio errors and then try this Button again. This Button WILL NOT appear in the CSV.',
+                  e,
+                )
+                resolve()
+              }
+
+              try {
+                // Add Twilio number to the Messaging Service
+                await twilioClient.messaging.services(TWILIO_MESSAGING_SERVICE_ID).phoneNumbers.create({ phoneNumberSid })
+                console.log(`  ... Added ${phoneNumber} to messaging service`)
+              } catch (e) {
+                console.error(
+                  '  ... ERROR adding the new Twilio phone number the messaging service, ***MUST*** do this manually in the Twilio UI. This Button WILL appear in the CSV.',
+                  e,
+                )
+              }
+            } else {
+              phoneNumber = phoneNumberOrAreaCode
+            }
+
+            // Write to CSV
+            csvWriter.write([uuid, response.unit, phoneNumber, serialNumber])
+
             resolve()
           })
         } else {